update reworked bottomNaviBar added hiding the password added modal window of delete/exit of account added deleting of user's review

+++ app/src/main/java/com/example/teashop/screen/screen/profile_screen/profile/ProfileScreen.kt
@@ -12,8 +12,12 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.LaunchedEffect
@@ -45,6 +49,7 @@ import com.example.teashop.screen.screen.main_screen.BottomSheetBonuses
 import com.example.teashop.ui.theme.Black10
 import com.example.teashop.ui.theme.Green10
 import com.example.teashop.ui.theme.Grey10
+import com.example.teashop.ui.theme.Red10
 import com.example.teashop.ui.theme.TeaShopTheme
 import com.example.teashop.ui.theme.White10
 import com.example.teashop.ui.theme.montserratFamily
@@ -83,10 +88,23 @@ fun LaunchProfileScreen(navController: NavController){
 
 var orderCount = ""
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun MakeProfileScreen(user: User?, tokenStorage: TokenStorage, context: Context, navController: NavController){
     orderCount = user?.ordersCount.toString()
     var bonusInfo by remember {
+        mutableStateOf(false)
+    }
+    var expandedExit by remember{
+        mutableStateOf(false)
+    }
+    var expandedDelete by remember{
+        mutableStateOf(false)
+    }
+    var exitAccount by remember{
+        mutableStateOf(false)
+    }
+    var deleteAccount by remember{
         mutableStateOf(false)
     }
 
@@ -203,22 +221,47 @@ fun MakeProfileScreen(user: User?, tokenStorage: TokenStorage, context: Context,
             icon = R.drawable.exit_icon,
             title = "Выйти из аккаунта",
             onClick = {
-                tokenStorage.deleteToken(context)
-                navController.navigate(
-                    Screen.Log.route,
-                    navOptions = navOptions {
-                        popUpTo(navController.graph.id) {
-                            inclusive = true
-                        }
-                    }
-                )
-            })
-        ProfileCard(icon = R.drawable.delete_icon, title = "Удалить аккаунт")
+                expandedExit = true
+            }
+        )
+        ProfileCard(
+            icon = R.drawable.delete_icon,
+            title = "Удалить аккаунт",
+            onClick = {
+                expandedDelete = true
+            }
+        )
     }
     if(bonusInfo){
         BottomSheetBonuses(header = "Что такое бонусы?", textId = R.string.BonusInfo) {
             bonusInfo = it
         }
+    }
+    if(expandedExit){
+        ConfirmButton(
+            header = "Выйти из аккаунта?",
+            accountAction = { exitAccount = it },
+            expandedChange = { expandedExit = it }
+        )
+    }
+    if(expandedDelete){
+        ConfirmButton(
+            header = "Удалить аккаунт?",
+            accountAction = { deleteAccount = it },
+            expandedChange = { expandedDelete = it }
+        )
+    }
+    if(exitAccount){
+        tokenStorage.deleteToken(context)
+        navController.navigate(
+            Screen.Log.route,
+            navOptions = navOptions {
+                popUpTo(navController.graph.id) {
+                    inclusive = true
+                }
+            }
+        )
+        exitAccount = false
     }
 }
 
@@ -271,9 +314,69 @@ fun ProfileCard(icon: Int, title: String, counter: Boolean = false, onClick: () 
     }
 }
 
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ConfirmButton(header: String, accountAction: (Boolean) -> Unit = {}, expandedChange: (Boolean) -> Unit = {}){
+    AlertDialog(onDismissRequest = { expandedChange(false) }) {
+        Card(
+            colors = CardDefaults.cardColors(containerColor = White10)
+        ) {
+            Column(
+                modifier = Modifier
+                    .padding(top = 10.dp)
+                    .fillMaxWidth(),
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                Text(
+                    text = header,
+                    fontFamily = montserratFamily,
+                    fontSize = 15.sp,
+                    fontWeight = FontWeight.W500,
+                    color = Black10
+                )
+                Row(
+                    modifier = Modifier.padding(10.dp)
+                ) {
+                    Button(
+                        onClick = {accountAction(true)},
+                        colors = ButtonDefaults.buttonColors(containerColor = Red10),
+                        modifier = Modifier
+                            .padding(end = 10.dp)
+                            .weight(1f)
+                    ) {
+                        Text(
+                            text = "Да",
+                            fontFamily = montserratFamily,
+                            fontSize = 15.sp,
+                            fontWeight = FontWeight.W500,
+                            color = White10
+                        )
+                    }
+                    Button(
+                        onClick = { expandedChange(false) },
+                        colors = ButtonDefaults.buttonColors(containerColor = Green10),
+                        modifier = Modifier
+                            .padding(start = 10.dp)
+                            .weight(1f)
+                    ) {
+                        Text(
+                            text = "Нет",
+                            fontFamily = montserratFamily,
+                            fontSize = 15.sp,
+                            fontWeight = FontWeight.W500,
+                            color = White10
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
+
 @Preview(showBackground = true)
 @Composable
 fun GreetingPreview() {
     TeaShopTheme {
+        ConfirmButton(header = "Выйти из аккаунта?")
     }
 }
